GameRules theory method refactor, updated README
