+++ README.md
@@ -1,2 +1,31 @@
-# PPS_VikingChess
+# PPS_VikingChess 1.0
 Hnefatafl Game development with Scala-Java-Prolog integration.
+
+You can play hnefatafl in 4 variants in pvp offline and pve:
+
+- hnefatafl 11x11
+- tawlbwrdd 11x11
+- tablut 9x9
+- brandubh 9x9
+
+With multiple sandwich captures, different dispositions and different ways to take the white king.
+
+You can open the in game menu whenever you want during a match and select what to do:
+- continue the match
+- restart the match
+- leave the match and start another one (or close the game)
+
+You can navigate through all moves of the match and undo the last move whenever you want during a match.
+
+****** NEW FOR VERSION 1.0 ******
+
+Now you can play against different mighty IA levels in pve:
+- newcomer: very easy
+- standard: normal difficulty
+- advanced: quite hard
+
+Team Members:
+- Celli Marco: marco.celli7@studio.unibo.it
+- Leo Pasquale: pasquale.leo@studio.unibo.it
+- Nannini Luca: luca.nannini2@studio.unibo.it
+- Speciale Giovanni Maria: giovanni.speciale@studio.unibo.it
