+++ src/main/scala/model/prolog/GameRules.scala
@@ -1,15 +1,16 @@
 package model.prolog
-import org.apache.commons.io.IOUtils
-import java.io.StringWriter
+import java.io.{BufferedReader,InputStreamReader}
+import java.nio.charset.StandardCharsets
+import java.util.stream.Collectors
 
 object GameRules {
 
   val THEORY: String = "/prolog/gameRules.pl"
 
   def theory(): String = {
-    val writer = new StringWriter
-    IOUtils.copy(getClass.getResourceAsStream(THEORY), writer, "utf8")
-    writer.toString
+    new BufferedReader(new InputStreamReader(getClass.getResourceAsStream(THEORY), StandardCharsets.UTF_8))
+    .lines()
+    .collect(Collectors.joining("\n"))
   }
 
 }
