+++ src/main/scala/view/factories/FontProvider.scala
@@ -2,6 +2,8 @@ package view.factories
 
 import java.awt.Font
 
+import view.utils.ResourceLoader
+
 object FontProvider {
 
   val HNEFATAFL_FONT_PATH = "/font/NorseBold-2Kge.otf"
